% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bccp.R
\name{bccp-package}
\alias{bccp-package}
\title{Bin-Conditional Conformal Prediction}
\description{
Implements bin-conditional conformal prediction for generating
valid prediction intervals with coverage guarantees within user-specified
outcome bins.
}
